created new file structure and organized
