--- animals.py
@@ -1,212 +1,0 @@
-# import the python datetime module to help us create a timestamp
-from datetime import date
-
-class MallardsFish:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.swimming = True
-
-Aimee = MallardsFish("Miss Aimee", "MarllardsFish")
-# Aimee.name = "Miss Aimee"
-# Aimee.species = "MarllardsFish"
-print(Aimee)
-
-class Goldfish:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.swimming = True
-Nat = Goldfish("Mr Nat", "Goldfish")
-# Nat.name = "Mr Nat"
-# Nat.species = "Goldfish"
-print(Nat)
-
-class KoiFish:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.swimming = True
-Cameron = KoiFish("Miss Cameron", "Fish")
-# Cameron.name = "Miss Cameron"
-# Cameron.species = "Fish"
-print(Cameron)
-
-class CarpsFish:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.swimming = True
-Elizabeth = CarpsFish("Miss Elizabeth", "Fish" )
-# Elizabeth.name = "Miss Elizabeth"
-# Elizabeth.species = "Fish"
-print(Elizabeth)
-
-class IdeFish:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.swimming = True
-Jack = IdeFish("Mr Jack", "Fish")
-# Jack.name = "Mr Jack"
-# Jack.species = "Fish"
-print(Jack)
-
-class CopperheadsSnake:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.slithering = True
-Juliet = CopperheadsSnake("Miss Juliet", "Snake" )
-# Juliet.name = "Miss Juliet"
-# Juliet.species = "Snake"
-print(Juliet)
-
-class RatSnake:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.slithering = True
-Yvee = RatSnake("Miss Yvee", "Snake")
-# Yvee.name = "Miss Yvee"
-# Yvee.species = "Snake"
-print(Yvee)
-
-class HognoseSnake:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.slithering = True
-Grace = HognoseSnake("Miss Grace", "Snake" )
-# Grace.name = "Miss Grace"
-# Grace.species = "Snake"
-print(Grace)
-
-class BallPythonSnake:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.slithering = True
-Charlie = BallPythonSnake("Mr Charlie", "snake" )
-# Charlie.name = "Mr Charlie"
-# Charlie.species = "snake"
-print(Charlie)
-
-class MilkSnake:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.slithering = True
-Huw = MilkSnake("Mr Huw", "domestic snake")
-# Huw.name = "Mr Huw"
-# Huw.species = "domestic snake"
-print(Huw)
-
-class Donkey:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.walking = True
-Paul = Donkey("Mr Paul", "Donkey" )
-# Paul.name = "Mr Paul"
-# Paul.species = "Donkey"
-print(Paul.name)
-
-class Llama:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.walking = True
-Jess = Llama("Miss Fuzz", "domestic llama" )
-# Jess.name = "Miss Fuzz"
-# Jess.species = "domestic llama"
-print(Jess)
-
-class Goat:
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.walking = True
-Lucy = Goat("Lucy", "goat" )
-# Lucy.name = "Lucy"
-# Lucy.species = "goat"
-print(Lucy)
-print(Lucy.name)
-
-class Goose:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.walking = True
-Oscar = Goose("Mr Oscar", "domestic goose" )
-# Oscar.name = "Mr Oscar"
-# Oscar.species = "domestic goose"
-print(Oscar)
-
-class Pig:
-
-    def __init__(self, name, species):
-        # Establish the properties of each animal
-        # with a default value
-        self.name = ""
-        self.species = ""
-        self.date_added = date.today()
-        self.walking = True
-Alienor = Pig("Mr Alienor", "Pig" )
-# Alienor.name = "Mr Alienor"
-# Alienor.species = "Pig"
-print(Alienor)
